+++ flask/models/introduction.py
@@ -7,7 +7,7 @@ from dataclasses import dataclass, asdict, fields
 from uuid import uuid4
 from typing import Optional, Any, List
 
-from database import Database
+from common.database import Database
 
 
 class IntroductionFields:
@@ -64,7 +64,7 @@ class Introduction:
         Args:
             server (Any): The CouchDB server instance.
         """
-        self.database = Database(server)
+        self.database = Database(server, "documents")
 
     def get(self, query_filter: dict) -> List[IntroductionModel]:
         """
